# docs/openapi.yaml
openapi: 3.0.3
info:
  title: thun.ai Backend API
  version: 1.0.0
  description: Machine-readable contract for the mobile, backend, privacy, and feedback flows.
servers:
  - url: https://api.thun.ai
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PrivacyConsentRequest:
      type: object
      required: [telemetryUpload, biometricsProcessing, therapistTranscriptRetention]
      properties:
        consentVersion:
          type: string
        telemetryUpload:
          type: boolean
        biometricsProcessing:
          type: boolean
        therapistTranscriptRetention:
          type: boolean
        marketingUpdates:
          type: boolean
paths:
...
  /privacy/consent:
    get:
      security:
        - bearerAuth: []
      summary: Get current consent state and recent privacy requests
      responses:
        '200':
          description: Privacy consent state
    put:
      security:
        - bearerAuth: []
      summary: Update privacy consent state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyConsentRequest'
      responses:
        '200':
          description: Updated consent
  /privacy/export:
    post:
      security:
        - bearerAuth: []
      summary: Queue a data export request
      responses:
        '202':
          description: Export request queued
  /privacy/delete-account:
    post:
      security:
        - bearerAuth: []
      summary: Queue an account deletion request
      responses:
        '202':
          description: Deletion request queued